requirements.txt
pyyaml
ipaddress